Begin documenting API endpoints.
Start documentation for API endpoints.  This commit includes relatively
complete documentation for the Client resource, and much sketchier
information for the Enrollments resource.  Next will be adding the
remaining the subresources under Enrollments, and adding examples for
all of those.  The Organizations and Projects resources are not yet
started.

Note that I've committed this on the feature-compass_schema branch
because recent API development has been happening on that branch, and
that is the branch that I've been testing on.  CONTRIBUTING.md specifies
that documentation specifically related to work done on a feature branch
should be pushed to that branch.